python tests: Skip python tests on Windows where POSIX pid is used
There is a difference between POSIX pid and Windows pid, not all the time are equal.
On Windows when a python script is started, a sh command is triggered as the parent
for script. So when we try to get the daemon pid with 'echo $!', this will get the pid of sh
not of python.exe as expected.
Also the 'kill' command expects a POSIX pid, not the Windows pid written by the python
daemons in pid file.

Signed-off-by: Paul-Daniel Boca <pboca@cloudbasesolutions.com>