python tests: Ported UNIX sockets to Windows
AF_UNIX sockets are not supported on Windows.
Instead of an AF_UNIX socket use localhost tcp connections to communicate
between components. This makes the python sockets compatible with
the ones used in Windows applications.

In case the socket returns WSAEWOULDBLOCK, it is replaced by EAGAIN error
in order to be compatible with Windows.

Signed-off-by: Paul-Daniel Boca <pboca@cloudbasesolutions.com>